Population Files now in separate folder + refactoring

+++ ExpandingGA/FileHandling/FileCreator.cs
@@ -18,11 +18,11 @@ namespace GeneticAlgorithmForStrings {
 
 		internal FileCreator(int generation, int individual, Individual genes) {
 
-			_dllDirectoryPath = System.IO.Path.Combine(FolderName, "DLL");
-			_directoryPath = System.IO.Path.Combine(FolderName, "Robots_gen" + generation.ToString("D4"));
+			_dllDirectoryPath = Path.Combine(FolderName, "DLL");
+			_directoryPath = Path.Combine(FolderName, "Robots_gen" + generation.ToString("D4"));
 
-			System.IO.Directory.CreateDirectory(_directoryPath);
-			System.IO.Directory.CreateDirectory(_dllDirectoryPath);
+			Directory.CreateDirectory(_directoryPath);
+			Directory.CreateDirectory(_dllDirectoryPath);
 
 			CreateFiles(generation, individual, genes);
 		}
@@ -34,26 +34,32 @@ namespace GeneticAlgorithmForStrings {
 			BattleFileCreator.CreateBattleFiles(_directoryPath, NameSpace, GetRobotName(generation, individual));
 		}
 
-		internal static void CreateFile(string filePath, string name, string contents) {
-			var pathIncludingFile = System.IO.Path.Combine(filePath, name);
+		internal static void CreateFile(string filePath, string name, string contents, bool overwrite) {
+			var pathIncludingFile = Path.Combine(filePath, name);
 
-			/** If we don't want to overwrite, the code changes a bit
-			if (!File.Exists(pathIncludingFile)) {
+
+			if (!overwrite)
+			{
+				if (!File.Exists(pathIncludingFile))
+				{
 					Console.WriteLine($"File \"{name}\" created!");
 					File.WriteAllText(pathIncludingFile, contents);
 				}
-				else {
+				else
+				{
 					Console.WriteLine($"File \"{name}\" already exists. Did not overwrite");
 				}
-			*/
-
-			Console.WriteLine(!System.IO.File.Exists(pathIncludingFile) ? "File \"{0}\" Created!" : "File \"{0}\" overwritten!", filePath + name);
-
-			// Create and write to file.
-			using (var fs = File.Create(pathIncludingFile)) {
-				var info = new UTF8Encoding(true).GetBytes(contents);
-				fs.Write(info, 0, info.Length);
-				fs.Close();
+			}
+			else
+			{
+				Console.WriteLine(!File.Exists(pathIncludingFile) ? "File \"{0}\" Created!" : "File \"{0}\" overwritten!", filePath + "/" + name);
+				// Create and write to file.
+				using (var fs = File.Create(pathIncludingFile))
+				{
+					var info = new UTF8Encoding(true).GetBytes(contents);
+					fs.Write(info, 0, info.Length);
+					fs.Close();
+				}
 			}
 		}
